refactor(agent): extract voice turn engine
voice-do.ts had grown to 896 lines, mixing the pure STT→LLM→TTS turn
engine with the hibernatable-WebSocket transport shell. Extract the turn
engine — frame/option types, runVoiceTurn, and the pure audio/identity
helpers (toByteIterable, pcmToWav, parseIdentity, readTranscriptionText,
readSynthesisAudio) — into voice-turn.ts, leaving voice-do.ts as the DO
transport only (473 lines). Public surface is unchanged: index.ts still
re-exports VoiceSessionDO (named) plus the turn types and runVoiceTurn.
Rename the turn test to voice-turn.test.ts to match its unit.
